feat: consolidate packages into frames.js
