+++ examples/framesjs-starter/app/examples/new-api/frames/frames.ts
@@ -1,4 +1,4 @@
-import { createFrames } from "@frames.js/next";
+import { createFrames } from "frames.js/next";
 
 // @todo here is issue with exporting, ts2742, how to fix it?
 export const frames = createFrames({
