+++ examples/framesjs-starter/app/examples/new-api/frames/route.tsx
@@ -1,5 +1,5 @@
 /* eslint-disable react/jsx-key */
-import { Button } from "@frames.js/next";
+import { Button } from "frames.js/next";
 import { frames } from "./frames";
 
 const handleRequest = frames(async ({ pressedButton, message }) => {
