+++ examples/framesjs-starter/app/examples/new-api/page.tsx
@@ -2,7 +2,7 @@ import Link from "next/link";
 import { currentURL } from "../../utils";
 import { createDebugUrl } from "../../debug";
 import type { Metadata } from "next";
-import { fetchMetadata } from "@frames.js/next";
+import { fetchMetadata } from "frames.js/next";
 
 export async function generateMetadata(): Promise<Metadata> {
   return {
