+++ examples/framesjs-starter/package.json
@@ -20,7 +20,6 @@
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.0",
     "frames.js": "^0.8.4",
-    "@frames.js/next": "^0.0.0",
     "lucide-react": "^0.331.0",
     "next": "14.1.0",
     "react": "^18.2.0",
