--- packages/express/tsconfig.json
@@ -1,5 +1,0 @@
-{
-  "extends": "@framesjs/typescript-config/nextjs.json",
-  "include": ["src"],
-  "exclude": ["dist", "build", "node_modules"],
-}
