+++ packages/frames.js/package.json
@@ -181,6 +181,56 @@
         "types": "./dist/core/index.d.cts",
         "default": "./dist/core/index.cjs"
       }
+    },
+    "./hono": {
+      "import": {
+        "types": "./dist/hono/index.d.ts",
+        "default": "./dist/hono/index.js"
+      },
+      "require": {
+        "types": "./dist/hono/index.d.cts",
+        "default": "./dist/hono/index.cjs"
+      }
+    },
+    "./remix": {
+      "import": {
+        "types": "./dist/remix/index.d.ts",
+        "default": "./dist/remix/index.js"
+      },
+      "require": {
+        "types": "./dist/remix/index.d.cts",
+        "default": "./dist/remix/index.cjs"
+      }
+    },
+    "./next": {
+      "import": {
+        "types": "./dist/next/index.d.ts",
+        "default": "./dist/next/index.js"
+      },
+      "require": {
+        "types": "./dist/next/index.d.cts",
+        "default": "./dist/next/index.cjs"
+      }
+    },
+    "./next/pages-router": {
+      "import": {
+        "types": "./dist/next/pages-router.d.ts",
+        "default": "./dist/next/pages-router.js"
+      },
+      "require": {
+        "types": "./dist/next/pages-router.d.cts",
+        "default": "./dist/next/pages-router.cjs"
+      }
+    },
+    "./express": {
+      "import": {
+        "types": "./dist/express/index.d.ts",
+        "default": "./dist/express/index.js"
+      },
+      "require": {
+        "types": "./dist/express/index.d.cts",
+        "default": "./dist/express/index.cjs"
+      }
     }
   },
   "files": [
@@ -189,9 +239,14 @@
   ],
   "devDependencies": {
     "@open-frames/types": "^0.0.6",
+    "@remix-run/node": "^2.8.1",
+    "@types/express": "^4.17.21",
     "@xmtp/frames-client": "^0.4.3",
     "@xmtp/frames-validator": "^0.5.2",
     "@xmtp/xmtp-js": "^11.5.0",
+    "express": "^4.19.1",
+    "hono": "^4.1.3",
+    "supertest": "^6.3.4",
     "tsup": "^8.0.1",
     "typescript": "^5.3.3"
   },
