--- packages/hono/jest.config.js
@@ -1,21 +1,0 @@
-/** @type {import('jest').Config} */
-export default {
-  testEnvironment: "node",
-  transform: {
-    ".*\\.(tsx?|jsx?)$": [
-      "@swc/jest",
-      {
-        jsc: {
-          transform: {
-            react: {
-              runtime: "automatic",
-            },
-          },
-        },
-      },
-    ],
-  },
-  moduleFileExtensions: ["ts", "tsx", "js", "jsx"],
-  transformIgnorePatterns: [],
-  testPathIgnorePatterns: ["/dist/", "/node_modules/"],
-};
