--- packages/next/package.json
@@ -1,70 +1,0 @@
-{
-  "name": "@frames.js/next",
-  "description": "Use frames.js in your Next.js app",
-  "version": "0.0.0",
-  "type": "module",
-  "main": "./dist/index.cjs",
-  "types": "index.d.cts",
-  "typesVersions": {
-    "*": {
-      "*": [
-        "./dist/*"
-      ]
-    }
-  },
-  "scripts": {
-    "lint": "eslint \"./src/**/*.ts*\"",
-    "build": "tsup `find ./src -type f \\( -regex '.*\\.tsx*' -a -not -regex '.*\\.test\\.tsx*' \\)` --format cjs,esm --dts",
-    "dev": "npm run build -- --watch",
-    "test:watch": "jest --watch"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/framesjs/frames.js/tree/main/packages/next"
-  },
-  "exports": {
-    ".": {
-      "import": {
-        "types": "./dist/index.d.ts",
-        "default": "./dist/index.js"
-      },
-      "require": {
-        "types": "./dist/index.d.cts",
-        "default": "./dist/index.cjs"
-      }
-    },
-    "./pages-router": {
-      "import": {
-        "types": "./dist/pages-router.d.ts",
-        "default": "./dist/pages-router.js"
-      },
-      "require": {
-        "types": "./dist/pages-router.d.cts",
-        "default": "./dist/pages-router.cjs"
-      }
-    }
-  },
-  "files": [
-    "dist",
-    "src"
-  ],
-  "engines": {
-    "node": ">=18"
-  },
-  "engineStrict": true,
-  "devDependencies": {
-    "@xmtp/frames-validator": "^0.5.1",
-    "tsup": "^8.0.1",
-    "typescript": "^5.3.3"
-  },
-  "license": "MIT",
-  "peerDependencies": {
-    "@xmtp/frames-validator": "^0.5.1",
-    "next": "14.1.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
-  },
-  "dependencies": {
-    "frames.js": "^0.8.4"
-  }
-}
